refactor(routes): relocate routed pages into src/pages/
Move HomePage, AuthPage, AppPage, MessagesPage, and RoomPage out of
src/prototypes/ and into src/pages/ so the prototypes folder is
reserved for reference material only. Replaces the older stub copies
that previously lived in src/pages/. App.tsx imports updated to point
at the new locations; src/prototypes/ is now empty and removed.
