+++ src/App.tsx
@@ -1,9 +1,9 @@
 import { BrowserRouter, Route, Routes, useParams } from 'react-router-dom'
-import HomePage from './prototypes/HomePage'
-import AuthPage from './prototypes/AuthPage'
-import AppPage from './prototypes/AppPage'
-import MessagesPage from './prototypes/MessagesPage'
-import RoomPage from './prototypes/RoomPage'
+import HomePage from './pages/HomePage'
+import AuthPage from './pages/AuthPage'
+import AppPage from './pages/AppPage'
+import MessagesPage from './pages/MessagesPage'
+import RoomPage from './pages/RoomPage'
 import AuthCallback from './pages/AuthCallback'
 import AppShell from './layouts/AppShell'
